requirements.txt
pdfminer.six==20220524
python-docx==1.1.2
scikit-learn==1.3.2
transformers==4.38.2   # updated
fastapi==0.95.0
pandas==2.1.3
torch==2.2.2
